moved ProcessScheduler to own file, separate from Process
